requirements.txt
art
rich.console
rich.panel
rich.syntax
rich.markdown
bs4